Update structure for Fuel
**Is this a new addition or an update to an existing factor?**

- [x] Update

**What category or categories does it affect?**

Fuel

**What?**

Updates all current Fuel IDs to adhere to the following structure:

fuel_type [e.g. "butane"]-activtity[e.g. "fuel_combustion"]

In addition to IDs, names and descriptions are also updated.

**Why?**

This is part of a larger refactoring of IDs across the OEFDB to adhere to a specified structure per category, allowing them to be better understood both by humans and machines.